Refactored model imports
