+++ backend/dao/database.py
@@ -1,30 +1,26 @@
 import psycopg2
 import os
 
-host = os.getenv('DB_HOST')
-user = os.getenv('DB_USER')
-password = os.getenv('DB_PASSWORD')
-database = os.getenv('DB_NAME')
+HOST = os.getenv('DB_HOST')
+USER = os.getenv('DB_USER')
+PASSWORD = os.getenv('DB_PASSWORD')
+DATABASE = os.getenv('DB_NAME')
+
+CONNECTION_STRING = f"host={HOST} user={USER} dbname={DATABASE} password={PASSWORD}"
 
 
-def execute_query(query: str) -> None:  # Function for use with insert and create query
-    connection_string = f"host={host} user={user} dbname={database} password={password}"
-    connection = psycopg2.connect(connection_string)
-    cursor = connection.cursor()
-    cursor.execute(query)
-    connection.commit()
-    cursor.close()
-    connection.close()
+def execute_query(query: str) -> None:
+    with psycopg2.connect(CONNECTION_STRING) as connection:
+        cursor = connection.cursor()
+        cursor.execute(query)
+        connection.commit()
 
 
-def select_query(query: str) -> list:  # Function for return values (Use with select)
-    connection_string = f"host={host} user={user} dbname={database} password={password}"
-    connection = psycopg2.connect(connection_string)
-    cursor = connection.cursor()
-    cursor.execute(query)
-    result = cursor.fetchall()
-    connection.commit()
-    cursor.close()
-    connection.close()
+def select_query(query: str) -> list:
+    with psycopg2.connect(CONNECTION_STRING) as connection:
+        cursor = connection.cursor()
+        cursor.execute(query)
+        result = cursor.fetchall()
+        connection.commit()
 
     return result
